backend/requirements.txt
aiosqlite==0.20.*
alembic==1.13.*
Authlib==1.3.0
fastapi==0.111.*
httpx==0.27.*
ipython==8.24.0
itsdangerous==2.2.0
Jinja2==3.1.*
asyncpg==0.29.*
pytest==8.2.*
requests==2.32.*
SQLAlchemy==2.0.*
uvicorn==0.29.*
bcrypt==4.1.*
sqlalchemy-utils==0.41.*
python-multipart==0.0.*
pyjwt==2.8.*
python-dateutil==2.9.*
pydantic-settings==2.2.*
pydantic[email]==2.7.*
lxml==4.9.3
